chore: move files for folders and make some changes
