+++ src/use-cases/errors/pacient/pacient-already-exists-error.ts
@@ -1,0 +1,5 @@
+export class PacientAlreadyExistsError extends Error {
+  constructor() {
+    super("Pacient e-mail already exists!");
+  }
+}
