--- src/use-cases/errors/user-already-exists-error.ts
@@ -1,5 +1,0 @@
-export class UserAlreadyExistsError extends Error {
-    constructor() {
-        super("E-mail already exists!");
-    }
-}
