+++ src/use-cases/factories/clinic/make-clinic-use-case.ts
@@ -1,0 +1,12 @@
+import { PrismaClinicsRepository } from "@/repositories/prisma/prisma-clinics-repository";
+import { RegisterClinicUseCase } from "../../modules/clinic/registerClinic";
+
+export const makeClinicUseCase = () => {
+  const prismaClinicsRepository = new PrismaClinicsRepository();
+
+  const registerClinicUseCase = new RegisterClinicUseCase(
+    prismaClinicsRepository,
+  );
+
+  return registerClinicUseCase;
+};
